requirements.txt
absl-py==2.1.0
accelerate==0.34.2
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
bert-score==0.3.13
bitsandbytes==0.45.1
black==24.10.0
blinker==1.9.0
Brotli==1.0.9
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.3.2
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
datasets==3.2.0
decord==0.6.0
deepspeed==0.16.3
detectron2==0.6
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.6.1
einops-exts==0.0.4
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.112.2
ffmpy==0.5.0
filelock==3.13.1
flash-attn==2.7.3
fonttools==4.55.6
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2==2.1.2
gradio==4.39.0
gradio_client==1.1.1
grpcio==1.70.0
h11==0.14.0
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
iopath==0.1.9
Jinja2==3.1.4
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.7
latex2mathml==3.77.0
linkify-it-py==2.0.3
Markdown==3.7
markdown-it-py==2.2.0
markdown2==2.5.3
MarkupSafe==2.1.3
matplotlib==3.9.4
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mmcls==0.25.0
mmsegmentation==0.30.0
model-index==0.1.11
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
narwhals==1.23.0
networkx==3.2.1
ninja==1.11.1.3
nltk==3.9.1
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
openai==1.63.2
opencv-python==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.0.11
ordered-set==4.1.0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pathspec==0.12.1
peft==0.10.0
pillow==10.4.0
pip==24.2
platformdirs==4.3.6
portalocker==3.1.1
prettytable==3.13.0
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
py-cpuinfo==9.0.0
pyarrow==19.0.0
pycocoevalcap==1.2
pycocotools==2.0.8
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rouge_score==0.1.2
rpds-py==0.22.3
ruff==0.9.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.13.1
segment_anything==1.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==2.20.0
setproctitle==1.3.4
setuptools==75.1.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
starlette==0.38.6
streamlit==1.41.1
streamlit-image-select==0.6.0
svgwrite==1.4.3
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
timm==0.9.12
tokenizers==0.15.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
transformers==4.37.2
triton==3.1.0
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
urllib3==2.3.0
uvicorn==0.34.0
wandb==0.19.4
watchdog==6.0.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
Werkzeug==3.1.3
wheel==0.44.0
xxhash==3.5.0
yacs==0.1.8
yapf==0.40.1
yarl==1.18.3
zipp==3.21.0
